@@ -1,16 +1,16 @@
 name: overflow_view
 description: A widget displaying children in a line with an overflow indicator at the end if there is not enough space.
-version: 0.3.1
+version: 0.4.0
 homepage: https://github.com/letsar/overflow_view
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.24.0-10.2.pre"
+  sdk: ">=2.17.0 <4.0.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  value_layout_builder: ^0.3.1
+  value_layout_builder: ^0.4.0
 
 dev_dependencies:
   flutter_test:
